pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "video-transcriber"
version = "0.1.0"
description = "Extract, transcribe, and intelligently organize audio from videos on YouTube, TikTok, or Instagram"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "ferlopezcarr" }
]
keywords = ["transcription", "video", "ai", "summarization", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "yt-dlp>=2024.4.9",
    "faster-whisper>=0.10.0",
    "openai-whisper>=20230314",
    "whisper>=1.1.10",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/ferlopezcarr/video_transcriber"
Repository = "https://github.com/ferlopezcarr/video_transcriber"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
target-version = "py312"
line-length = 100
